convert: split the emitters into focused modules
Pure moves, no logic changes: identifiers and the folding family,
partitions, tables, constraints, indexes, views, and the sequence,
synonym and db-link extras each get their own module, with Residue
in a small shared one. schema.py shrinks to the orchestrator and
keeps re-exports so existing imports hold. Same estate converts to
byte-identical output before and after.
