+++ src/pgrecon/convert/tables.py
@@ -1,0 +1,158 @@
+"""Table emission: typed columns, defaults, and partition children."""
+
+import sqlite3
+
+from pgrecon.convert.identifiers import _default_guard, _fold_expression, ident
+from pgrecon.convert.partitions import _emit_partition_children, _partition_meta
+from pgrecon.convert.residue import Residue
+from pgrecon.convert.typemap import map_type
+
+
+def _columns(conn: sqlite3.Connection, owner: str, table: str) -> list[sqlite3.Row]:
+    return conn.execute(
+        "SELECT column_name, data_type, data_length, data_precision,"
+        " data_scale, nullable FROM columns"
+        " WHERE owner = ? AND table_name = ? ORDER BY position",
+        (owner, table),
+    ).fetchall()
+
+
+def emit_tables(
+    conn: sqlite3.Connection,
+    out: list[str],
+    residue: list[Residue],
+    emitted: dict[tuple[str, str], set[str]],
+    dropped: dict[tuple[str, str], set[str]],
+) -> tuple[int, int]:
+    """Emit every table; returns (table_count, partition_count)."""
+    table_count = 0
+    partition_count = 0
+
+    tables = conn.execute(
+        "SELECT owner, table_name, temporary FROM tables ORDER BY owner, table_name"
+    ).fetchall()
+
+    for t in tables:
+        owner, table = t["owner"], t["table_name"]
+        cols = _columns(conn, owner, table)
+        if not cols:
+            residue.append(
+                Residue(owner, table, "table", "no column facts in the inventory")
+            )
+            continue
+        extras = {
+            (r["column_name"] or "").upper(): r
+            for r in conn.execute(
+                "SELECT column_name, default_text, virtual, truncated"
+                " FROM column_defaults WHERE owner = ? AND table_name = ?",
+                (owner, table),
+            )
+        }
+        lines: list[str] = []
+        kept_columns: set[str] = set()
+        for c in cols:
+            mapped = map_type(
+                c["data_type"], c["data_length"], c["data_precision"], c["data_scale"]
+            )
+            if mapped.pg_type is None:
+                residue.append(
+                    Residue(
+                        owner,
+                        f"{table}.{c['column_name']}",
+                        "column",
+                        mapped.note or "unmappable type",
+                    )
+                )
+                dropped.setdefault((owner, table), set()).add(
+                    (c["column_name"] or "").upper()
+                )
+                continue
+            if mapped.note is not None:
+                residue.append(
+                    Residue(owner, f"{table}.{c['column_name']}", "note", mapped.note)
+                )
+            cname = (c["column_name"] or "").upper()
+            null = "" if (c["nullable"] or "Y") == "Y" else " NOT NULL"
+            suffix = ""
+            extra = extras.get(cname)
+            if extra is not None and (extra["virtual"] or "NO") == "YES":
+                expr = (
+                    None
+                    if extra["truncated"]
+                    else _fold_expression(extra["default_text"] or "")
+                )
+                guard = None if expr is None else _default_guard(expr)
+                if expr is None or guard is not None:
+                    residue.append(
+                        Residue(
+                            owner,
+                            f"{table}.{c['column_name']}",
+                            "column",
+                            guard
+                            or "virtual column expression could not be"
+                            " translated; recreate it by hand",
+                        )
+                    )
+                    dropped.setdefault((owner, table), set()).add(cname)
+                    continue
+                suffix = f" GENERATED ALWAYS AS ({expr}) STORED"
+            elif extra is not None and (extra["default_text"] or "").strip():
+                if extra["truncated"]:
+                    residue.append(
+                        Residue(
+                            owner,
+                            f"{table}.{c['column_name']}",
+                            "note",
+                            "default was truncated during extraction;"
+                            " column emitted without it",
+                        )
+                    )
+                else:
+                    folded = _fold_expression(extra["default_text"])
+                    guard = None if folded is None else _default_guard(folded)
+                    if folded is None or guard is not None:
+                        residue.append(
+                            Residue(
+                                owner,
+                                f"{table}.{c['column_name']}",
+                                "note",
+                                (guard or "default could not be translated")
+                                + "; column emitted without it",
+                            )
+                        )
+                    else:
+                        suffix = f" DEFAULT {folded}"
+            lines.append(
+                f"    {ident(c['column_name'])} {mapped.pg_type}{suffix}{null}"
+            )
+            kept_columns.add(cname)
+        if not lines:
+            residue.append(
+                Residue(owner, table, "table", "every column was unconvertible")
+            )
+            continue
+        meta, part_reason = _partition_meta(conn, owner, table)
+        if part_reason is not None:
+            residue.append(Residue(owner, table, "partitioning", part_reason))
+        if (t["temporary"] or "N") == "Y":
+            residue.append(
+                Residue(
+                    owner,
+                    table,
+                    "table",
+                    "global temporary table emitted as a plain table;"
+                    " per-session semantics need pgtt or a redesign",
+                )
+            )
+        out.append(f"CREATE TABLE {ident(table)} (")
+        out.append(",\n".join(lines))
+        out.append(f"){meta.clause if meta else ''};")
+        out.append("")
+        table_count += 1
+        emitted[(owner, table)] = kept_columns
+        if meta is not None:
+            partition_count += _emit_partition_children(
+                conn, owner, table, meta, out, residue
+            )
+
+    return table_count, partition_count
